TST: split up test_concat.py #37243
 * created test_index.py

 * small changes to test_empty.py, test_datetimes.py, test_dataframe.py
