+++ pandas/tests/reshape/concat/test_datetimes.py
@@ -518,3 +518,15 @@ class TestPeriodConcat:
         result = concat([x, y], ignore_index=True)
         tm.assert_series_equal(result, expected)
         assert result.dtype == "object"
+
+
+def test_concat_timedelta64_block():
+    from pandas import to_timedelta
+
+    rng = to_timedelta(np.arange(10), unit="s")
+
+    df = DataFrame({"time": rng})
+
+    result = concat([df, df])
+    assert (result.iloc[:10]["time"] == rng).all()
+    assert (result.iloc[10:]["time"] == rng).all()
